+++ system/MDR32F9Qx/MDR_config.c
@@ -10,13 +10,13 @@ __IO uint64_t _micros = 0;
 UART_BUFFER_INDEX_T _uart1_rx_buffer_head = 0;
 //указывает на следуюую ячеку для чтения
 UART_BUFFER_INDEX_T _uart1_rx_buffer_tail = 0;
-short _uart1_rx_buffer[SERIAL_RX_BUFFER_LENGTH];
+uint8_t _uart1_rx_buffer[SERIAL_RX_BUFFER_LENGTH];
 
 //указывает на следуюую ячеку для записи
 UART_BUFFER_INDEX_T _uart2_rx_buffer_head = 0;
 //указывает на следуюую ячеку для чтения
 UART_BUFFER_INDEX_T _uart2_rx_buffer_tail = 0;
-short _uart2_rx_buffer[SERIAL_RX_BUFFER_LENGTH];
+uint8_t _uart2_rx_buffer[SERIAL_RX_BUFFER_LENGTH];
 
 
 //CAN
@@ -29,4 +29,4 @@ CAN_RX_Package _can_rx_buffer[CAN_RX_BUFFER_LENGTH];
  //указывает на следуюую ячеку для чтения
  NRF24_BUFFER_INDEX_T _nrf24_rx_buffer_tail = 0;
 
- short _nrf24_rx_buffer[NRF24_RX_BUFFER_LENGTH];
+ uint8_t _nrf24_rx_buffer[NRF24_RX_BUFFER_LENGTH];
